# Controllers/CNCProgramController.cs
using Microsoft.AspNetCore.Mvc;
using System.Collections.Generic;
using System.Threading.Tasks;
using HavenCNCServer.Models;

namespace HavenCNCServer.Controllers
{
    /// <summary>
    /// CNC Program Control - Handles G-code execution, step run, and program control
    /// </summary>
    [ApiController]
    [Route("api/[controller]")]
    public class CNCProgramController : ControllerBase
    {
        #region G-Code Execution Control

        /// <summary>
        /// Stop G-code execution
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("Stop")]
        public async Task<IActionResult> Stop()
        {
            // TODO: Implement stop functionality
            await Task.Delay(1);
            return Ok(new { message = "Execution stopped" });
        }

        /// <summary>
        /// Resume G-code execution
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("Resume")]
        public async Task<IActionResult> Resume()
        {
            // TODO: Implement resume functionality
            await Task.Delay(1);
            return Ok(new { message = "Execution resumed" });
        }

        /// <summary>
        /// Resume G-code execution at specific line
        /// </summary>
        /// <param name="lineNumber">Line number to resume at</param>
        /// <returns>Success response</returns>
        [HttpPost("ResumeAt/{lineNumber}")]
        public async Task<IActionResult> ResumeAt(int lineNumber)
        {
            // TODO: Implement resume at line functionality
            await Task.Delay(1);
            return Ok(new { message = $"Execution resumed at line {lineNumber}", lineNumber });
        }

        /// <summary>
        /// Run G-code
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("RunGCode")]
        public async Task<IActionResult> RunGCode()
        {
            // TODO: Implement run G-code functionality
            await Task.Delay(1);
            return Ok(new { message = "G-code execution started" });
        }

        /// <summary>
        /// Run single G-code command
        /// </summary>
        /// <param name="gcode">G-code command to run</param>
        /// <returns>Success response</returns>
        [HttpPost("RunGCodeCommand")]
        public async Task<IActionResult> RunGCodeCommand([FromBody] string gcode)
        {
            // TODO: Implement run G-code command functionality
            await Task.Delay(1);
            return Ok(new { message = $"Executed G-code: {gcode}", gcode });
        }

        #endregion

        #region G-Code Management

        /// <summary>
        /// Get current G-code
        /// </summary>
        /// <returns>Current G-code lines</returns>
        [HttpGet("GetCurrentGCode")]
        public async Task<IActionResult> GetCurrentGCode()
        {
            // TODO: Implement get current G-code
            await Task.Delay(1);
            var gcode = new List<string> { "G00 X0 Y0", "G01 Z-1 F100", "M03 S1000" };
            return Ok(new { gcode });
        }

        /// <summary>
        /// Load G-code
        /// </summary>
        /// <param name="request">G-code load request</param>
        /// <returns>Success response</returns>
        [HttpPost("LoadGCode")]
        public async Task<IActionResult> LoadGCode([FromBody] LoadGCodeRequest request)
        {
            // TODO: Implement load G-code functionality
            await Task.Delay(1);
            return Ok(new { message = $"Loaded {request.GCode.Count} G-code lines", lineCount = request.GCode.Count });
        }

        /// <summary>
        /// Load G-code from file
        /// </summary>
        /// <param name="filePath">File path to load G-code from</param>
        /// <returns>Success response</returns>
        [HttpPost("LoadGCodeFromFile")]
        public async Task<IActionResult> LoadGCodeFromFile([FromBody] string filePath)
        {
            // TODO: Implement load G-code from file functionality
            await Task.Delay(1);
            return Ok(new { message = $"G-code loaded from {filePath}", filePath });
        }

        /// <summary>
        /// Get current line number
        /// </summary>
        /// <returns>Current line number</returns>
        [HttpGet("GetCurrentLineNumber")]
        public async Task<IActionResult> GetCurrentLineNumber()
        {
            // TODO: Implement get current line number
            await Task.Delay(1);
            return Ok(new { lineNumber = 1 });
        }

        #endregion

        #region Step Run Control

        /// <summary>
        /// Start step run mode
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("StartStepRun")]
        public async Task<IActionResult> StartStepRun()
        {
            // TODO: Implement start step run functionality
            await Task.Delay(1);
            return Ok(new { message = "Step run mode started" });
        }

        /// <summary>
        /// End step run mode
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("EndStepRun")]
        public async Task<IActionResult> EndStepRun()
        {
            // TODO: Implement end step run functionality
            await Task.Delay(1);
            return Ok(new { message = "Step run mode ended" });
        }

        /// <summary>
        /// Execute next step in step run mode
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("StepRunNext")]
        public async Task<IActionResult> StepRunNext()
        {
            // TODO: Implement step run next functionality
            await Task.Delay(1);
            return Ok(new { message = "Next step executed" });
        }

        /// <summary>
        /// Run from current step
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("RunFromCurrentStep")]
        public async Task<IActionResult> RunFromCurrentStep()
        {
            // TODO: Implement run from current step functionality
            await Task.Delay(1);
            return Ok(new { message = "Running from current step" });
        }

        #endregion
    }
}