# Controllers/CNCSpindleController.cs
using Microsoft.AspNetCore.Mvc;
using System.Threading.Tasks;
using HavenCNCServer.Models;

namespace HavenCNCServer.Controllers
{
    /// <summary>
    /// CNC Spindle Control - Handles all spindle operations and speed control
    /// </summary>
    [ApiController]
    [Route("api/[controller]")]
    public class CNCSpindleController : ControllerBase
    {
        #region Spindle Control

        /// <summary>
        /// Start the spindle
        /// </summary>
        /// <param name="speed">Optional speed parameter</param>
        /// <returns>Success response</returns>
        [HttpPost("StartSpindle")]
        public async Task<IActionResult> StartSpindle([FromBody] double? speed = null)
        {
            // TODO: Implement start spindle
            await Task.Delay(1);
            string message = speed.HasValue ? $"Spindle started at {speed}" : "Spindle started";
            return Ok(new { message, speed });
        }

        /// <summary>
        /// Stop the spindle
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("StopSpindle")]
        public async Task<IActionResult> StopSpindle()
        {
            // TODO: Implement stop spindle
            await Task.Delay(1);
            return Ok(new { message = "Spindle stopped" });
        }

        /// <summary>
        /// Warm up the spindle
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("WarmUpSpindle")]
        public async Task<IActionResult> WarmUpSpindle()
        {
            // TODO: Implement spindle warm up
            await Task.Delay(1);
            return Ok(new { message = "Spindle warm up initiated" });
        }

        /// <summary>
        /// Check if spindle is running
        /// </summary>
        /// <returns>Spindle running status</returns>
        [HttpGet("IsSpindleRunning")]
        public IActionResult IsSpindleRunning()
        {
            // TODO: Implement spindle running check
            return Ok(new { isRunning = false });
        }

        #endregion

        #region Spindle Speed Control

        /// <summary>
        /// Get current spindle speed
        /// </summary>
        /// <returns>Current spindle speed</returns>
        [HttpGet("GetSpindleSpeed")]
        public async Task<IActionResult> GetSpindleSpeed()
        {
            // TODO: Implement get spindle speed
            await Task.Delay(1);
            return Ok(new { speed = 1000 });
        }

        /// <summary>
        /// Set spindle speed
        /// </summary>
        /// <param name="speed">Speed to set</param>
        /// <returns>Success response</returns>
        [HttpPost("SetSpindleSpeed")]
        public async Task<IActionResult> SetSpindleSpeed([FromBody] double speed)
        {
            // TODO: Implement set spindle speed
            await Task.Delay(1);
            return Ok(new { message = $"Spindle speed set to {speed}", speed });
        }

        /// <summary>
        /// Adjust spindle speed by factor
        /// </summary>
        /// <param name="factor">Adjustment factor (-200 to +200, where 0 is no change, 100 is double, -100 is half)</param>
        /// <returns>Success response</returns>
        [HttpPost("AdjustSpindleSpeed")]
        public async Task<IActionResult> AdjustSpindleSpeed([FromBody] double factor)
        {
            // TODO: Implement adjust spindle speed
            await Task.Delay(1);
            return Ok(new { message = $"Spindle speed adjusted by factor {factor}", factor });
        }

        /// <summary>
        /// Get current spindle speed adjustment factor
        /// </summary>
        /// <returns>Current adjustment factor</returns>
        [HttpGet("GetCurrentSpindleSpeedFactor")]
        public async Task<IActionResult> GetCurrentSpindleSpeedFactor()
        {
            // TODO: Implement get current spindle speed factor
            await Task.Delay(1);
            return Ok(new { factor = 0.0 });
        }

        /// <summary>
        /// Reset spindle speed factor to default
        /// </summary>
        /// <returns>Success response</returns>
        [HttpPost("ResetSpindleSpeedFactor")]
        public async Task<IActionResult> ResetSpindleSpeedFactor()
        {
            // TODO: Implement reset spindle speed factor
            await Task.Delay(1);
            return Ok(new { message = "Spindle speed factor reset to default" });
        }

        #endregion
    }
}