# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HavenCNC Server API",
    "description": "REST API for CNC server management hosted within WinForms application",
    "contact": {
      "name": "HavenCNC Server Team"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/CNCProgram/ResumeAt/{lineNumber}": {
      "post": {
        "tags": [
          "CNCProgram"
        ],
        "summary": "Resume G-code execution at specific line",
        "parameters": [
          {
            "name": "lineNumber",
            "in": "path",
            "description": "Line number to resume at",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/CNCProgram/RunGCodeCommand": {
      "post": {
        "tags": [
          "CNCProgram"
        ],
        "summary": "Run single G-code command",
        "requestBody": {
          "description": "G-code command to run",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/CNCProgram/LoadGCode": {
      "post": {
        "tags": [
          "CNCProgram"
        ],
        "summary": "Load G-code",
        "requestBody": {
          "description": "G-code load request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadGCodeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadGCodeRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoadGCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CNCProgram/LoadGCodeFromFile": {
      "post": {
        "tags": [
          "CNCProgram"
        ],
        "summary": "Load G-code from file",
        "requestBody": {
          "description": "File path to load G-code from",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/CNCSpindle/StartSpindle": {
      "post": {
        "tags": [
          "CNCSpindle"
        ],
        "summary": "Start the spindle",
        "requestBody": {
          "description": "Optional speed parameter",
          "content": {
            "application/json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            },
            "text/json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/CNCSpindle/SetSpindleSpeed": {
      "post": {
        "tags": [
          "CNCSpindle"
        ],
        "summary": "Set spindle speed",
        "requestBody": {
          "description": "Speed to set",
          "content": {
            "application/json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            },
            "text/json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CNCSpindle/AdjustSpindleSpeed": {
      "post": {
        "tags": [
          "CNCSpindle"
        ],
        "summary": "Adjust spindle speed by factor",
        "requestBody": {
          "description": "Adjustment factor (-200 to +200, where 0 is no change, 100 is double, -100 is half)",
          "content": {
            "application/json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            },
            "text/json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "number",
                "format": "double"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoadGCodeRequest": {
        "type": "object",
        "properties": {
          "gCode": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "G-code lines to load",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Request model for loading G-code"
      },
...
    }
  }
}